Add files via upload
Hey, So I don't know what I'm doing.  I made a profile (Probably based of of CHEP, but I don't remember).   This Profile seems to be going well using the following:
1. Nylon 230 - https://www.matterhackers.com/store/3d-printer-filament/taulman-nylon-230-filament--1.75mm (Non affiliate link)
2.  Ender 3 Pro with Sprite Extruder (https://amzn.to/3KjUpS8) , Linear Rails, Auto Bed Leveler CR Touch (https://amzn.to/3KjUpS8) , Dual Z Axis.

Certainly there are things that can be improved, but this is a good start for someone looking (possibly).

I recommend running A calibration cube and making fine adjustments after.

All links above are affiliate links which means i may make a commission based upon the sale of a product.  If you don't want to use those links, it's fine, enjoy the profile.

https://www.youtube.com/c/hamradiodude if you want to support by subscribing.   I'm long winded.